requirements.txt
admitad==1.1.1
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
future==1.0.0
h11==0.14.0
idna==3.5
numpy==1.24.4
openpyxl==3.1.2
outcome==1.3.0.post0
pandas==2.0.3
psycopg2-binary==2.9.9
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tabulate==0.9.0
trio==0.23.1
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.1.0
wsproto==1.2.0
